- added parameter "useThermalNoise" to BasePhyLayer - added a BaseDecider implementation and derived SNRThresholdDecider and Decider80211 from it - added "@unit" to parameters of BasePhyLayer and BaseMacLayer - fixed a bug in "MappingUtils::findMin/Max"
git-svn-id: https://mixim.svn.sourceforge.net/svnroot/mixim/trunk@499 ccb5d51d-3e1d-0410-9886-bf3c9037c363

+++ base/phyLayer/Interpolation.h
@@ -209,7 +209,7 @@ public:
 	static double linearInterpolation(const Key& t,
 									  const Key& t0, const Key& t1,
 									  const V& v0, const V& v1){
-		return v0 + (v1 - v0) * (t - t0) / (t1 - t0);
+		return v0 + (v1 - v0) * (double)((t - t0) / (t1 - t0));
 	}
 
 	interpolated operator()(const InputIterator& first,
